refactor: 메서드를 MoneyController, NumberController로 분리
